pyproject.toml
[tool.poetry]
name = "Test"
version = "0.1.0"
description = "Test"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = ">=2.0.0"
nonebot-adapter-onebot = ">=2.2.1"
requests = "^2.22.0"
pymysql = "^0.9.3"
python-office = "^0.3.19"
playwright = "1.30.0"
imagehash = "4.2.1"
matplotlib = "3.3.3"
"ruamel.yaml" = "^0.17.10"
httpx = "^0.18.2"
nonebot_plugin_apscheduler = "^0.1.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
